+++ lib/definitions/project.d.ts
@@ -444,6 +444,10 @@ interface IAssetItem {
 	resizeOperation?: string;
 	overlayImageScale?: number;
 	rgba?: boolean;
+
+	// additional operations for special cases
+	operation?: "delete" | "writeXMLColor";
+	data?: any;
 }
 
 interface IAssetSubGroup {
@@ -475,6 +479,7 @@ interface IImageDefinitionGroup {
 interface IImageDefinitionsStructure {
 	ios: IImageDefinitionGroup;
 	android: IImageDefinitionGroup;
+	android_legacy: IImageDefinitionGroup;
 }
 
 interface ITemplateData {