+++ lib/services/android-plugin-build-service.ts
@@ -144,6 +144,40 @@ export class AndroidPluginBuildService implements IAndroidPluginBuildService {
 		return promise;
 	}
 
+	private getIncludeGradleCompileDependenciesScope(
+		includeGradleFileContent: string
+	): Array<string> {
+		const indexOfDependenciesScope = includeGradleFileContent.indexOf(
+			"dependencies"
+		);
+		const result: Array<string> = [];
+
+		if (indexOfDependenciesScope === -1) {
+			return result;
+		}
+
+		const indexOfRepositoriesScope = includeGradleFileContent.indexOf(
+			"repositories"
+		);
+
+		let repositoriesScope = "";
+		if (indexOfRepositoriesScope >= 0) {
+			repositoriesScope = this.getScope(
+				"repositories",
+				includeGradleFileContent
+			);
+			result.push(repositoriesScope);
+		}
+
+		const dependenciesScope = this.getScope(
+			"dependencies",
+			includeGradleFileContent
+		);
+		result.push(dependenciesScope);
+
+		return result;
+	}
+
 	private getScope(scopeName: string, content: string): string {
 		const indexOfScopeName = content.indexOf(scopeName);
 		const openingBracket = "{";
@@ -375,14 +409,19 @@ export class AndroidPluginBuildService implements IAndroidPluginBuildService {
 		const allGradleTemplateFiles = path.join(gradleTemplatePath, "*");
 		const buildGradlePath = path.join(pluginTempDir, "build.gradle");
 		const settingsGradlePath = path.join(pluginTempDir, "settings.gradle");
-		
+
 		this.$fs.copyFile(allGradleTemplateFiles, pluginTempDir);
+		this.addCompileDependencies(platformsAndroidDirPath, buildGradlePath);
 		const runtimeGradleVersions = await this.getRuntimeGradleVersions(
 			projectDir
 		);
 		this.replaceGradleVersion(
 			pluginTempDir,
 			runtimeGradleVersions.gradleVersion
+		);
+		this.replaceGradleAndroidPluginVersion(
+			buildGradlePath,
+			runtimeGradleVersions.gradleAndroidPluginVersion
 		);
 		this.replaceFileContent(buildGradlePath, "{{pluginName}}", pluginName);
 		this.replaceFileContent(settingsGradlePath, "{{pluginName}}", pluginName);
@@ -430,12 +469,13 @@ export class AndroidPluginBuildService implements IAndroidPluginBuildService {
 		let runtimeVersion: string = null;
 
 		try {
-			const result = await this.$packageManager.view(
+			let result = await this.$packageManager.view(
 				SCOPED_ANDROID_RUNTIME_NAME,
 				{
 					"dist-tags": true,
 				}
 			);
+			result = result?.["dist-tags"] ?? result;
 			runtimeVersion = result.latest;
 		} catch (err) {
 			this.$logger.trace(
@@ -529,6 +569,7 @@ export class AndroidPluginBuildService implements IAndroidPluginBuildService {
 				`${SCOPED_ANDROID_RUNTIME_NAME}@${runtimeVersion}`,
 				{ version_info: true }
 			);
+			output = output?.["version_info"] ?? output;
 
 			if (!output) {
 				/**
@@ -543,6 +584,7 @@ export class AndroidPluginBuildService implements IAndroidPluginBuildService {
 					`${SCOPED_ANDROID_RUNTIME_NAME}@${runtimeVersion}`,
 					{ gradle: true }
 				);
+				output = output?.["gradle"] ?? output;
 
 				const { version, android } = output;
 
@@ -603,6 +645,22 @@ export class AndroidPluginBuildService implements IAndroidPluginBuildService {
 		);
 	}
 
+	private replaceGradleAndroidPluginVersion(
+		buildGradlePath: string,
+		version: string
+	): void {
+		const gradleAndroidPluginVersionPlaceholder =
+			"{{runtimeAndroidPluginVersion}}";
+		const gradleAndroidPluginVersion =
+			version || AndroidBuildDefaults.GradleAndroidPluginVersion;
+
+		this.replaceFileContent(
+			buildGradlePath,
+			gradleAndroidPluginVersionPlaceholder,
+			gradleAndroidPluginVersion
+		);
+	}
+
 	private replaceFileContent(
 		filePath: string,
 		content: string,
@@ -612,6 +670,29 @@ export class AndroidPluginBuildService implements IAndroidPluginBuildService {
 		const contentRegex = new RegExp(content, "g");
 		const replacedFileContent = fileContent.replace(contentRegex, replacement);
 		this.$fs.writeFile(filePath, replacedFileContent);
+	}
+
+	private addCompileDependencies(
+		platformsAndroidDirPath: string,
+		buildGradlePath: string
+	): void {
+		const includeGradlePath = path.join(
+			platformsAndroidDirPath,
+			INCLUDE_GRADLE_NAME
+		);
+		if (this.$fs.exists(includeGradlePath)) {
+			const includeGradleContent = this.$fs.readText(includeGradlePath);
+			const compileDependencies = this.getIncludeGradleCompileDependenciesScope(
+				includeGradleContent
+			);
+
+			if (compileDependencies.length) {
+				this.$fs.appendFile(
+					buildGradlePath,
+					"\n" + compileDependencies.join("\n")
+				);
+			}
+		}
 	}
 
 	private copyAar(
@@ -709,7 +790,9 @@ export class AndroidPluginBuildService implements IAndroidPluginBuildService {
 				projectDir: pluginBuildSettings.projectDir,
 			});
 			pluginBuildSettings.androidToolsInfo = this.$androidToolsInfo.getToolsInfo(
-				{ projectDir: pluginBuildSettings.projectDir }
+				{
+					projectDir: pluginBuildSettings.projectDir,
+				}
 			);
 		}